Better URL parsing for Kwai URLs. (#255)
We didn't have a specific parser for Kwai URLs, so it was falling back to oEmbed information and metatags, but it means that the title and description are the same for all shared URLs from Kwai. This change add a specific parser that is able to extract the user who shared the video and the user who created the video, and use that information to fill some of the Pender data fields.

Reference: CHECK-1806.