Replace tests against Meedan website
With the Meedan site redesign, we seem to no longer have social tags
in the header. As a result, the tests in Pender we use to validate
some parsing behavior (on author and opengraph metadata) is failing.

To address this for now, I went ahead and updated the tests to read
from other URLs that have the data we're looking for. We should
consider doing two things longer term:
1. Add social metadata to Meedan.com, and possibly add the Meedan.com
tests back in here or elsewhere, since they seem useful to tell us
about our own site even if it wasn't the specific intention of these
tests
2. Consider testing against fixtures instead of live websites