Update Instagram GraphQL URL for items
At some point recently the URL we use to query GraphQL data for
Instagram items stopped returning the data we need. This updates that
URL with (honestly kinda magical) query params that make the
query work again. It's possible this will change again in the future--
would be nice to find an approach that works and is documented.

Here's the StackOverflow post that suggested the query params
needed: https://stackoverflow.com/a/72582413

CHECK-1201