Provide encrypted cookies to Travis for test runs
In order to run our tests that reach out to Instagram, we need to provide
cookies in the request. Since they're sensitive, we need to store them as
an encrypted file that is decrypted as part of the CI run. This commit adopts
the approach taken for `config.yml` to encrypt the `cookies.txt`.

At the moment, only cookies for Instagram are provided.

For more on encrypted files on Travis, see: https://docs.travis-ci.com/user/encrypting-files/#Encrypting-multiple-files
Some things to note about that documentation:
* To log in and sometimes for various commands, you have to provide the --pro flag, since Travis CLI still defaults to .org
* Contrary to the documentation, encrypting multiple files is now supported

CHECK-1201