Fix Instagram profile parsing
The approach we were using to parse Instagram profiles broke because
Instagram no longer includes OpenGraph header information on profiles.
To get information for the profile, we now are using a private
Instagram API with a magical value for ig_app_id that mimics
making the request in the browser.

I worry this will still be brittle in the future, but this gets us going.
This also updates tests so that they're a little more black box, which I
hope makes changes easier in future - just set the API response and
check the values.

CHECK-1201