Send Instagram API-related errors to Errbit
Previously we were swallowing errors arising from requesting
and parsing Instagram-related requests. This changes the approach
to wrap what we think are errors relating from the Instagram API
and raise them to Errbit so that we can monitor those errors
for indications that the Instagram-related features are broken.
This is particularly important since we're using private APIs
that are finnicky.

The previous behavior of swallowing all other errors resulting
in the request is maintained since I wasn't sure of the ramifications
of changing it.

This sets up a pattern that we could use in future, as well -
each provider could have its own set of exceptions, which we then
send to Errbit or (if noisy) ignore in the Errbit config.

CHECK-1201