Delay initial metrics request
We were seeing a race condition where the metrics update
returned before the intiial request created the record on the
Check API-side. Since Check API doesn't currently return a
different status if the Media with URL isn't present in the
database, we want to just delay the initial call to try to
ensure that the record is created before we try to update it.