Simplify metrics-related methods
Since we only ever register one metrics provider (Facebook)
and don't right now intend to expand this, we can call the
Facebook-related metrics method directly and remove the
methods related to metrics-provider registration.

This also renames the methods to be more clear that we're
scheduling a background job, and to reduce the confusion between
two methods with the same name (one on instance level and one on
class level), though the pattern is a little different than
how we approach it for archivers.