Extract Metrics into its own class
Moves Metrics into its own class, which is passed its dependencies when
called from Media. This is the start of splitting out the media providers
into their own classes -- this was a bit simpler and a clear chunk of work,
so decided to start here first.

Also starts to move over Metrics-related tests to a new pattern of
mostly unit-testing (black-box style) without touching the internet,
while having one or two that still reach out to real sources to confirm
our isolated assumptions are correct. Also attempts to set up and teardown
in the tests as much as possible, instead of relying on config files.

Additionally sends all Facebook-related API errors to Errbit and logs,
so we can track them. Only retries on ones we had previously identified
as being retryable (in future we could add to this list if we want)