Skip metrics integration test if rate limited
We don't want to fail if the test comes back against expectation
because we're being rate limited. Also, we want to try as many
times as we reasonably can without firing off a ton of requests
that make the rate limit problem worse. This commit introduces a pattern
of retrying as many times as we have different facebook apps in the
test environment, and only testing against expectation if the request
returns sucessfully. If not, it notes the problem and skips the test.