Configure mocha to automatically teardown stubs after test run
In order to get mocha to tear down stubs, it has to be loaded last.
I think that previously we were importing it before some of our minitest
libraries, and also weren't using the minitest-specific version of the gem,
which meant that we lost the auto-teardown functionality.

Changing the version of mocha and the load order means we no longer need to
manually tear down mocks as part of test setup and teardown.
(See https://github.com/freerange/mocha#installation)

Also begins to remove unnecessary unstubbing at end of tests, and removes
some expectations that previously weren't running since they weren't accurate.