Turn off Sidekiq inline for Metrics unit tests
Right now in the test setup we force Sidekiq jobs to run inline (of
which there are many, since we queue followup jobs). I'd like to move
away from this so that we can have more control over our tests and mocks,
so I'm starting with the file we know is isolated.