Encapsulate request-related methods we will need for parsers
In order to break out the parsers into their own classes, we need to
be able to use certain methods that are right now within the Media class.
This commit extracts the request-related methods into a helper that
can be used by those classes, and also passes in dependent information
that we were reading off of Media where appropriate.

To reduce amount of breakage, this wraps the old methods in Media
around the new methods, so that we can begin to move over to the new
methods with potentially new method signatures at our own pace.