Begin converting parsers into classes, starting with KwaiItem
This was one of the simpler parsers, so began here with the move
to extract the parser modules from the Media class and into their
own isolated classes.

Kwai doesn't seem to use oembed at all, so that line was removed from
the previous implementation since it would require additional effort
to make the oembed parsing available to the separated class.