Convert TikTokProfile to parser class
This class is the first time we need to get updated HTML from a parser
class. As a result, we had to move the get_html function and make it
usable in the context of our new classes. We can keep most of the
previous state by passing an error-setting callback (since the existing
parser modules and new parser classes both have different expectations
of where errors will be set - and if we extract get_html to a separate
class then there's no "self.data" to set info on as there was before it
was split out of Media.

This also pulls over the set_data_field function, since we need the
ability to set specific fields only if they're not already set.