Convert TwitterItem to parser class
There are a number of tests that I'm not sure we need, but it's
hard to tell since they were all integration tests. I'm keeping
them around as commented for reference, and intend to take them
out of code base or update as I work through the other parsers -
particularly the oembed one.