Convert YoutubeItem to parser class
Also adds in a missing test about error handling in YoutubeProfile,
and consolidates some shared logic for the Youtube parsers.

There is some confusing logic around defaults when the YoutubeItem
fails to parse correctly (eg if the API response returns incorrectly)
but since this behavior already existed, I'm just porting it over
as-is for now.