Convert FacebookProfile parser to class
This is a fairly substantial re-write of the Facebook Profile
parser functionality. Much of the parsing logic seemed to no
longer work with the current HTML format of Facebook Pages.

This attempts to simplify the logic and rely less on HTML parsing,
which I expect will change and requires being able to successfully,
reach a page and more on information that we can grab from the URL.

It also attempts to reduce what seemed like duplicative network calls,
instead relying on the HTML doc passed from the media class.

Some notable changes:
* No longer fetches HTML from link, since it can be passed in as doc
* Does not reassign URL if URL is an ignored link
* Tries getting ID from URL based on patterns we use to idenitfy a match,
but just numeric
* Tries to parse ID from HTML using pageID - and no longer relying on
gem that was doing the same, since the logic to extract ID is simple and
gem is no longer maintained (last update 2016)
* Relies on HTML meta tags and things like title for information like
descrpition, picture URLs, and title, which seem to be fairly
consistent across different page markups
* Removes subtype, which was previously always returning 'page' since it
relied on parsing outdated markup. We also seem to no longer need this
information.

This intends to mimic the logic and information priority of the
returned values of the previous implementation