Extract Facebook Item Parser class
This is a significant re-write of the previous MediaFacebookItem class.
The main goal was to make it be a standalone class like the other
parsers that are being extracted from Media. To do this, a lot needed to
change. In doing this work, too, it became clear that much of our HTML
parsing logic was out of date.

This commit:
* Creates a new FacebookItem class, and moves much of the existing
behavior into the class
* Changes the parsing behavior to rely more heavily on Crowdtangle, instead
of HTML parsing (which was mostly broken, because of the frequency of changes)
and metatags (which we may add back in later). Instead, if a Crowdtangle request
fails (which it may often do - only a small subset of posts are available through
Crowdtangle) then we use the same fallback behavior as in other classes -
returning a unique title (set in the Media class) - and, if possible, returning
the HTML to be used for embed. The intention is for this to be simpler, require less
upkeep, and be less brittle than the HTML parsing that we had tried before.
* Moves UUID parsing into a separate class so that we can more easily
unit test the behavior, and pulls in the previous functinality from the
IdsPlease gem for easier maintainability. We seemed to only be using a
few regexes from the gem, which do work, but can be brought into our
code base here since we're no longer sharing the functionality and the
gem is not actively maintained (though it is ours).
* Enforces some new constraints for UUID format - before we were
allowing some alphabetical and single item IDs through, which Crowdtangle
does not support. This aims to be more strict, and rely more heavily on
regexes (vs parsing params and positional logic in URLs)
* Creates new unit tests to try to enforce what seems like existing
behavior (though hard to tell, given the range of differe Facebook items)