Extract PageItem parser and OembedItem classes
- Creates a new isolated OembedItem that can be used with Page Item
parser class without depending on internals of Media. Modifies MediaOembed
just enough to not require passing the oembed_url, which is now set
on the parser instance and not available in all contexts in which the MediaOembed
class method is called (rendering oembed HTML)
- Re-introduces opengraph and twitter metatag helpers for use in
the Page Item parser, which we could also use to make parsing more robust
in other classes. For now they are only in use in a few places, as
reflected in existing behavior.
- Introduces base oembed_url in the existing parser classes from the MediaOembedItem class,
so that we can use that method to send into the OembedItem class without
creating a dependency on the HTML doc. Parsers should know their oembed URL.
- Creates Page Item parsing class, reimplementing previous behavior but isolated
- Assigns nil values rather than empty strings in these classes where present, so that
it makes it easier to detect when a value is unset (can rely on truthiness
rather than having to call .blank?)