Remove encapsulated methods for RequestHelper
As part of the refactor to break parsers out into their own classes,
I had encapsulated some existing Media methods to reduce breakage
while I completed the move.

The parsing classes moves are now complete, so we can remove the
encapsulation of the RequestHelper methods and just use the methods
on RequestHelper directly.

For the most part it's just a rename. In at least one case, though -
absolute_url - the arguments were changed. In updating this, it revealed
a bug in the recently refactored OembedItem parser that removed support
for relative paths in Oembed URLs. This restores that support.