Fix oembed fetching behavior
OembedItem is now its own class that requires passing in an oembed_url,
which was done as part of the Parser class refactor. Because of that,
we now have to change the implementation of MediaOembed.as_oembed to
allow for fetching oembed data when there is a custom oembed
URL (for Facebook, TikTok, etc).

This now attempts to fetch oembed information as the parser dictates
(either from the custom URL or from the URL parsed from the HTML, if
a custom URL is not specified for the parser). If no valid oembed
data is present at those URLs, then it falls back to setting oembed
data from the Media object itself.