Allow setting Pender cookies from file on S3 (#286)
We're moving away from using Configurator, which means that we need
a way to include cookies that aligns with how we now treat other secrets.
This commit moves toward us setting an S3 file path in `config.yml`,
which would point to a file in an S3 bucket that contains the required
cookies. This request will be authenticated with the credentials
we already use for storage-related S3 items.

The config setting is `cookies_file_path`, and can point to the local
filesystem (for development and, if needed, test) or to a file on S3. 
If no `cookies_file_path` is provided in the config, it defaults to
what we previously used: `config/cookies.txt`.

This also:
* Removes cookies.txt from CI, since we would need a different way 
of providing cookies since we're retiring configurator. Also updates
tests to no longer rely on this text file locally, instead stubbing config.
* Fixes some typos in the config.example file.