Only check for pictures in at one URL in integration test
The Hacktoberfest pages recently changed (it's end of September)
and one of them was causing the integration test to fail, since
there was no longer an author photo. This commit changes the check
to just one URL, since we have more robust unit tests on the
parsing but could still use a check to make sure the data is
cached in PenderStore.