Normalize all media URLs
Previously we were only running canonical URLs through normalization
if they were not found in the HTML. This lead to an edge case where
we were returning different URLs for effectively the same item - one
if it was parsed in the HMTL (eg with a trailing slash) and a different
one if it used the passed URL (which was then run through
Postrank::URI.normalize, which stripped trailing /, among other things).

This is a pretty big change since it will mean that many duplicates may
be created in Check API (since it asks Pender for a canonical URL when
looking up existing Links, and that will change).