Prevent infinite loop if redirect chain never broken
We were seeing stack level too deep errors in tests on CI, possibly from
a hardcoded web mock response with a redirect. Cannot reproduce with
single test runs.

It's possible this would happen in a real situation, so I added a guard that prevents
the oembed method from recursively calling itself forever, similar to what
we already have in the original URL check in Media.