Add ability to sample traces in Open Telemetry
To start out, we want to send all information to Honeycomb. But, depending
on the volume of traces generated, we may want to reduce the amount of
traces that we send. This commit enables that: sending all data by default,
and setting a sampler with sample rate in the case that we want to limit.

To use, we would need to set the following in config.yml:
* `otel_traces_sampler` - should be a Open Telemetry supported value, like traceidratio
* `otel_custom_sampling_rate` - this should be a number, ideally an integer, representing
the amount we want to sample. It will be set as the relevant config
for Open Telemetry (e.g. if 5 if set, then SampleRate=5 will be sent as an attribute,
and OTEL_TRACES_SAMPLER_ARG will be set in environment to 0.2

This commit also consolidates instrumentation into one method that is
used across all environments, so that we have less of a chance of our test
and real environments diverging in enabled instrumentation and configuration.
Ideally we would move this full configuration to only run in *some* integration tests
in the future, but for now we run on everything to be safe.

In general, the complexity happening in Pender::OpenTelemetryConfig is attempting
to put some guard rails in configuring Open Telemetry, since there are many
pieces that need to be set to specific values for everything to run as expected.