Set defaults for unset Open Telemetry config
Since we load from yml file, we can easily pass nil in for these
values. This sets defaults assuming that sometimes we will be
passed nil, but we always want a hash.

It's possible we might want more guard rails in future (like,
erroring if the passed config doesn't have the shape we want, or
passing things in individually, etc) but this seems sufficient
to fix current issues with deployment.