Temporarily disable Tiktok shortlink test
The shortlink we use for integration tests has stopped resolving -
visiting it just redirects to the homepage, causing our tests to fail.
I haven't been able to figure out if that is Tiktok dropping support
for shortlinks (I can't seem to generate one with the same format, and
any URL I find online does the same thing). I'm marking this test as
skipped until we find out and either update the test or remove the
code to support shortlinks.