Fix flakiness in archiver tests
Previously we weren't calling up to the test_helper teardown, which meant
that the resetting of enabled archivers wasn't happening consistently. This
commit calls super to make sure we get both teardowns, and ensures that we
reset the values of the ENABLED_ARCHIVERS constant to the start of the test
when they change mid-test.