Remove remove_const to reduce test issues
We were using the private method remove_const in various places
because we're modifying a class constant in tests and otherwise
would get warnings when we re-defined.

Remove_const removed the definition of the class constant, which
caused problems when we tried to re-define it, since the constant
was actually defined in the MediaArchiver module but included
and generally accessed in Media. There was a lot going on. To get
around this complexity, I instead just redefined the constant
and temporarily suppressed the warning.

I think this is a bit of a code smell, and instead we might want to
move toward injecting dependecnies rather than constant modification.
But, I think this gets us going.