Give Check API a headstart before triggering archive jobs
We rely on Check API having created a link (inline) and archiver
annotation (background job) before we're able to update the media
with archiving information. While we are going to implement retry
behavior, we can also give Check API a bit more likelihood of
succeeding the first time by delaying the archiving job on our
side.