Raise exception when notify_webhook returns unsuccessfully
We were running into a number of race conditions with the endpoint where
our application would try to alert the endpoint about an interaction with
some external service (metrics, archiving) but Check API wasn't quite ready.
This would show up as Check API silently dropping the request and returning
successfully when the link/project media/annotation had not yet been created
to associate Pender's webhook request with.

We are intending to change this behavior, so that the Check API endpoint
returns an error indicating when a save has failed. This commit is preparing for
that, and raising an exception on an unsucesssful API response that will provoke
Sidekiq to retry again in the future.

Test update: Previously we were relying on an external endpoint existing in tests,
which was returning a 301. We now look at the return of that request
in order to figure out whether to raise an exception (any non-successful
response will cause this). This commit updates our tests to use the mocked
endpoint, which matches both the desired behavior and makes our tests
a bit more isolated from external dependencies.