Send metrics reporting error information to Honeycomb
Right now we get a lot of non-actionable errors relating to gathering and
reporting metrics. We currently send them to Errbit with unique names (caused
by including the 'message' key from Facebook, which includes things like an
object ID), which creates a lot of noise since we can't do anything about
the errors and they're all showing up as individual items instead of being grouped
by type.

So, this commit does a few things:
* Creates a new helper function to allow for more easily associating error data
with spans
* Makes the Facebook metrics-related errors have non-unique names when reported to
Errbit
* Associates error data with our Honeycomb spans both so we have more detailed data
and so that we do have some error data relating to metrics fetching even if we
suppress the failure (which we are now doing for retryable errors, though I don't
think those are the bulk of our metrics. Making names non-unique will help us tell)