Use normalized URL as cache key
Use the normalized URL when accessing cache

Previously we were using the original URL to control cache. This causes
a problem because we use the normalized URL to communicate with Check API,
causing a problem because there would often not be a cache present for
the same link that was previously submitted if the normalized verison was
different than the original link.

This takes the approach of using normalized URLs for caching, so we have a
single source of truth that is compatible with the way that Check API stores
URLs from Pender.

This also sends the normalized URL to the archiving service, instead of
the original URL. Using the original URL for archiving because the archiving
service wasn't matching the Link saved to Check API, which was stripped previously by Pender. This commit brings in line
the URL parsing behavior so that we can get a match our archive to the
same link.

Lastly, the change in caching behavior caused some tests to fail, since before
we could just generate a caache ID from whatever URL we set. Now we can only
use the normalized URL to generate an accurate cache key, which happens both
within Media.new and media.as_json. To sidestep this in tests we can either
use a normalized URL from the start (often easier, if we don't intend to test
normalization though can be confusing when it fails) or generate the cache ID
like so: `m = Media.new(url); id = Media.get_id(m.url)`