Prevent exponential creation of metrics jobs on failure
Previously, in some cases we were exponentially creating new Metrics updates
for URLs beacuse we queued the next update inside the existing background job,
and relied on a passed variable to dictate when to stop scheduling jobs (that
on error would not always be set).

This instead schedules all of our updates at the start (for ten days, as
was previously done) and caps the amount of retries up to one day (according
to the Sidekiq documentation for back-off timing).

In theory we could stop passing count / update_number, but this would change
the signature of the method, which would cause issues for currently queued
Sidekiq jobs, so I just left it and used it for some logging.