Create TelegramItem parser and set title to URL
Previously Telegram links were being parsed by the PageItem parser,
which used the non-unique title as our title, which is used for similarity
matching. Because Telegram sets the title of the page to the channel
that published it, this meant that items were being marked as similar
just because they were published by the same channel - not necessarily
the same topic.

To address this, this commit creates a TelegramItem parser that simplifies
the fetching logic (mostly the same as what PageItem did), and sets the
title as the URL since there is no other unique identifier provided (eg,
no post title).

Telegram profiles seemed to parse as I'd expect, so I continued to let
those fall back to PageItem as they do now.

CV2-2612