Update Ruby version to 2.7.7
Previous Ruby version reached end of life.

* Use Addressable gem instead of URI for URL parsing and encoding

To upgrade, most of the work was around no longer using URI.escape.
URI.escape is deprecated and will be removed in Ruby 3.0. In preparation
for upgrading, this commit moves to using the Addressable gem, which
implements relevant RFCs for URI encoding, decoding, and parsing.

https://www.rubydoc.info/gems/addressable/Addressable/URI

This commit uses the previous logic we had for where we should encode,
decode, and parse. Since Addressable::URI.parse encodes as part of
this process, we no longer need to encode before passing into the
parse method. As part of this, I changed references of .port to
.inferred_port, since the Addressable URI component only returns
 something from .port if it is set explicitly in the URL, which
it is usually not. Instead, .inferred_port looks to the HTTML
scheme and sets a port accordingly, which we want in most cases.

Also adds a method RequestHelper.encode_url so that we capture behavior in one
place, and renames .decoded_url -> .decode_url to make naming more
consistent across URL-management functions

* Strip params from Instagram profile URLs

This was unrelated to the Ruby upgrade, but began presentingn itself
during the upgrade so decided to fix.

Previously if a URL with a param made its way into the Instagram Profile
parser, the param would be included in the username / external ID
because the regex did not strip it. This was surfaced because one
of the external test URLs began returning params in the URL, presumably
from either the canonical URL or from a redirect.

This updates the regex to chop off anything after a question mark.

CV2-2668