Upgrade to Rail 6.0
Includes the following:
* Upgrade sqlite to be compatible with Rails version

We had an old version of sqlite pinned, seemingly due to a previous
incompatibility with Rails 5.2.2. This has since been fixed and we
need a new version of sqlite for Rails 6, so it gets unpinned and updated.

* Begin autoloading via zeitwork

In Rails 6, there is a new autoloader called zeitwerk. It is more
strict about where it expects constants to live - one class per file,
and each file should live in a directory structure that matches its
constant name (e.g. "Pender::Store" should be at "pender/store.rb", not
"pender_store.rb".

This commit begins using the Zeitwork autoloader by setting the following
in the application config: `config.load_defaults 6.0` and then updates
our lib constants to a directory structure that can be autoloaded. Since
we had an inconsistent module name, I also consolidated LapisConstants
into Lapis::, and created Pender::Exception::x as a bucket for all of our
Pender-related exceptions module, since they had to be split into their own
files but it made sense to give a way for them to be loaded as a set
(eg "pender/exception" by creating "pender/exception.rb").

The autoloading can be checked by running: `bin/rails zeitwerk:check`

More here: https://edgeguides.rubyonrails.org/upgrading_ruby_on_rails.html#upgrading-from-rails-5-2-to-rails-6-0

* Remove represent_boolean_as_integer = true, since it is now default and will be removed in 6.1

* Remove usage of ActionView::Base.new and replace with ApplicationController.render

This was causing errors because we were passing a Pathname of views to ActionView::Base.new,
and the constructor has changed to expect a lookup context. We were also receiving related
deprecation warnings (`DEPRECATION WARNING: ActionView::Base instances should be constructed
with a lookup context, assignments, and a controller.`)

We have two options to address this: passing in a lookup context to ActionView::Base, or
using ApplicationController.render. It seemed to me like we could use the latter and it
was a bit simpler of an interface, so I decided to move forward with that approach.

https://github.com/rails/rails/blob/v6.0.3.3/actionview/lib/action_view/base.rb#L258
https://stackoverflow.com/a/63865988

* Bundles URI-related errors so that we can more easily catch them

* Silences warnings from constant redefinition

Upgrades Bundler and explicitly declares Net HTTP gem so that
we stop getting so many warnings. Might be able to remove gem
declaration in Ruby 3.0.

References:
* https://github.com/ruby/net-imap/issues/16
* https://github.com/codeforamerica/vita-min/commit/835591e05f1fdb7b3f69613be1fd686932f30e44

* Add Spring to speed up development and tests

* Avoid autoloading constants during initialization

In Rails 6.0, autoloading constants during initialization has been deprecated.
This means that we should be explicitly requiring any constant that is used
in config/initializers rather than relying on autoload.

We were receiving warning signs about PenderConfig, ApiKey, and ApplicationRecord,
all of which were triggered by the inclusion of PenderConfig in the Airbrake
initializer. To address, I explicitly required each.

More information here:
* https://edgeguides.rubyonrails.org/upgrading_ruby_on_rails.html#upgrading-from-rails-5-2-to-rails-6-0
* https://bill.harding.blog/2021/07/22/rails-6-1-deprecation-warning-initialization-autoloaded-the-constants-what-to-do-about-it/

* Configure config.hosts in development to avoid DNS binding attacks

This is configured only in development based on recommendation of the
Rails guide, and to allow us to manage whitelisting via our infrastructure
in hosted environments.

More here:
* https://edgeguides.rubyonrails.org/upgrading_ruby_on_rails.html#new-config-hosts-setting
* https://github.com/rails/rails/pull/33145#issuecomment-398502787

* Use RequestHelper.request_url for all requests

In the Ruby 2.7 upgrade, we began using Addressable for URIs. This
resulted in some unexpected behavior when Net::HTTP.get_request was
used, beacuse Addressable::URI goes down a different code path than
URI because it is not recognized as a URI. THis was causing warnings
in Open Telemetry (because Addressable::URI made its way to target,
as opposed to the path) and potentially caused us to make incorrect
requests.

This standardizes the way we make requests across the application
when we don't need to set custom headers, using the non-shortcut
version of constructing Net::HTTP requests.

I'd like to further refactor this in the future - skipping unnecessary
lines (like teasing apart responsibilities of html_options, since we
often only use the User Agent) - and allowing for the pass-in of custom
headers, since we still use Net::HTTP to construct reqeusts elsewhere.

CV2-2668