Use postgres in development and test, instead of sqlite
Previously we were using sqlite for development and test environments
while our deployed environments used postgres. This caused an issue
during the Rails upgrade where our Postgres version was out of date
but never caught locally or in CI because of the difference in configuration.
This commit has us use postgres as the local database, removes references
to sqlite, and updates postgres to be compatible with the 6.1 version
of ActiveRecord.

CV2-2668