Fix failure to deploy due to missing storage.yml
After upgrading to Rails 6.1 and Ruby 3.0, our deploys began failing
due to a missing storage.yml. storage.yml is needed to configure ActiveStorage,
which we don't currently use.

To sidestep this, configure ActiveStorage to use :local in all environments.
We will want to change this in the future if we start using ActiveRecord.

CV2-2668