Add basic storage.yml
This is the default installed by Rails that we were missing due to
upgrades. Although we're not using ActiveStorage, our deploys are erroring
without this config file.

This also removes the declaration of local storage that had been previously
added to try to get around this same error, since I think we don't need it
and it will be confusing in the future.

CV2-2668