Use page title for certain Facebook item URL patterns
Facebook's item HTML markup is inconsistent for different URL patterns.
For most cases, using the og:description tag gets us the most unique and
accurate post contents. We began using this as the Facebook item title
in a previous commit. Unfortunately, for some URL patterns, using the
description gives us an incorrect result for the page title - usually,
contents of the event description instead of the name or for videos an
engagement summary. In these cases, we are making a carveout to instead
use the og:title or <title> tag. This should cover most cases, though
there are still a known few URL patterns this approach does not work for
(URLs with /videos/ in them, media sets - which will default back to the
URL for the title).

CV2-2737